+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## next
 * Checkbox - added styling and example
-* SearchBar replaced with @opuscapita/react-searchbar
+* SearchBar - replaced with @opuscapita/react-searchbar and added example
 
 ## 8.1.0
 * Upgrade - Update react icons 1.3.0